+++ organizations/src/resolvers/organization-resolver.ts
@@ -16,13 +16,13 @@ import { Context } from "../index";
 
 @Resolver((of) => Org)
 export class OrgResolver {
-  @Query(() => [Org])
+  @Query(() => [Org], { description: "list all organizations" })
   async list() {
     const orgs = await OrgModel.find({});
     return orgs;
   }
 
-  @Query(() => Org)
+  @Query(() => Org, { description: "fetch one organization by id" })
   async get(@Arg("id") id: string) {
     const org = await OrgModel.findById(id);
     if (!org) {
@@ -31,40 +31,17 @@ export class OrgResolver {
     return org;
   }
 
-  @Mutation(() => Org)
-  async add(@Arg("data") newOrg: AddOrgInput) {
+  @Mutation(() => Org, { description: "add new organization" })
+  async addOrg(@Arg("data") newOrg: AddOrgInput) {
     const org = OrgModel.create(newOrg);
     return org;
   }
 
-  @Mutation(() => Review)
-  async addReview(@Arg("data") newReview: AddReviewInput, @Ctx() ctx: Context) {
-    // auth check
-    if (!ctx.user) {
-      throw new NotAuthorizedError();
-    }
-    // step 1: find org
-    const org = await OrgModel.findById(newReview.organizationId);
-    if (!org) {
-      throw new NotFoundError();
-    }
-    // step 2: create review
-    const review = await ReviewModel.create({
-      ...newReview,
-      userId: ctx.user.id,
-      userEmail: ctx.user.email,
-    });
-
-    // step 3: update org
-    org.reviews?.push(review.id);
-    await org.save();
-    return review;
-  }
-
-  @FieldResolver()
+  @FieldResolver({
+    description: "fetch the reviews associated with an organization",
+  })
   async reviews(@Root("id") orgId: string) {
     const reviews = await ReviewModel.find({ organizationId: orgId });
-    console.log(reviews);
     return reviews;
   }
 }
